syscall: getsockopt SO_*TIMEO compared a signed socklen_t on Bionic
Building on 32-bit Termux (clang, Bionic) warned twice in the
SO_RCVTIMEO/SO_SNDTIMEO read-back of getsockopt:

  src/sys_net.c:580: comparison of integers of different signs:
      'socklen_t' (aka 'int') and 'unsigned int' [-Wsign-compare]
  src/sys_net.c:581: (same)

Bionic types socklen_t as a signed int where glibc and musl make it
unsigned, so the two `tl >= sizeof ...` tests -- how much of the host's
timeval the kernel actually filled in -- are a signed/unsigned
comparison only on Android and never on the development host, which
is why the warning surfaced there and nowhere else.

Cast the reported length to size_t for the comparison, the way the
address-length handling in the same file already does. On glibc/musl
the cast is a no-op; on Bionic a (never kernel-produced) negative
length now counts as huge rather than small, which is exactly what the
unsigned libcs have always done, so the three tiers agree.

The warning is reproducible on x86 without a device: a shim header
that re-includes glibc's bits/types.h and then `#define __socklen_t
int`, passed through `CC="cc -I<shimdir>"`, produced the same two
diagnostics on the old source and none on this one, and turned up no
other socklen_t sign comparison in the tree.

+++ src/sys_net.c
@@ -577,8 +577,8 @@ SYSDEF(getsockopt) {
         u8 g[16];
         memset(g, 0, sizeof g);
         s64 v;
-        if (tl >= sizeof tv.tv_sec) { v = (s64)tv.tv_sec; memcpy(g, &v, 8); }
-        if (tl >= sizeof tv) { v = (s64)tv.tv_usec; memcpy(g + 8, &v, 8); }
+        if ((size_t)tl >= sizeof tv.tv_sec) { v = (s64)tv.tv_sec; memcpy(g, &v, 8); }
+        if ((size_t)tl >= sizeof tv) { v = (s64)tv.tv_usec; memcpy(g + 8, &v, 8); }
         u32 outl = (u32)glen < 16 ? (u32)glen : 16;   /* kernel: len = min(len, lv) */
         if (outl && copy_to_guest(c, a3, g, outl) < 0) return (u64)(s64)-EFAULT;
         if (copy_to_guest(c, a4, &outl, 4) < 0) return (u64)(s64)-EFAULT;
